Refactor Gatus to use common base entity class
